backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# DB
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.9
alembic>=1.13.0

# LangChain / GROQ / LangGraph
langchain>=0.1.0
langchain-core>=0.1.0
langchain-groq>=0.0.6
langgraph>=0.0.20

# HTTP and data
httpx>=0.26.0
requests>=2.31.0
yfinance>=0.2.36
duckduckgo-search>=7.0.0
feedparser>=6.0.0

# Optional: sse-starlette for SSE
sse-starlette>=1.8.0
